fix(journal): split diskFreeBytes into build-tagged unix/windows files
unix.Statfs is not available on Windows; move the free-space probe behind
build constraints (unix uses statfs, windows uses GetDiskFreeSpaceEx) so the
Windows build compiles.
