+++ pkg/block/journal/store.go
@@ -11,8 +11,6 @@ import (
 	"sync/atomic"
 	"time"
 
-	"golang.org/x/sys/unix"
-
 	"github.com/marmos91/dittofs/internal/logger"
 	"github.com/marmos91/dittofs/pkg/block/chunker"
 )
@@ -335,17 +333,6 @@ func (s *Store) startBackgroundGC() {
 			}
 		}
 	}()
-}
-
-// diskFreeBytes reports the bytes available to an unprivileged writer on the
-// filesystem backing dir. It sizes Open's default local-store cap so an
-// unconfigured store still bounds its on-disk growth.
-func diskFreeBytes(dir string) (uint64, error) {
-	var st unix.Statfs_t
-	if err := unix.Statfs(dir, &st); err != nil {
-		return 0, err
-	}
-	return st.Bavail * uint64(st.Bsize), nil
 }
 
 // WriteAt buffers a dirty client write. It never fsyncs; durability is a
